+++ common/src/main/java/com/github/leopoko/solclassic/item/WickerBasketItem.java
@@ -115,6 +115,7 @@ public class WickerBasketItem extends Item {
 
     @Override
     public void appendHoverText(@NotNull ItemStack stack, @Nullable Level level, List<Component> tooltip, @NotNull TooltipFlag flag) {
+        if (!SolclassicConfigData.enableItemDescription) return;
         tooltip.add(Component.translatable("tooltip.wicker_basket.description1"));
         tooltip.add(Component.translatable("tooltip.wicker_basket.description2"));
     }